Login & Registration page
Summary
Added login page and registration page and fixed css style to match and fit into your style.css. I still need to fix the positioning of a button on my login and registration page as it is off by a little but it should be fine for now as it is only off by a little.

Added
- login.html
- register.html

updated
- login button to href link on all pages
- added to css style to match your style and my login page. There is no more overlapping in styles.